Refactor the TransactionHelper
