+++ core/src/main/java/org/stellar/anchor/util/TransactionHelper.java
@@ -1,38 +1,24 @@
-package org.stellar.anchor.platform.utils;
+package org.stellar.anchor.util;
 
-import static org.stellar.anchor.api.platform.PlatformTransactionData.Kind.*;
+import static org.stellar.anchor.api.platform.PlatformTransactionData.Kind.RECEIVE;
 
 import javax.annotation.Nullable;
-import lombok.SneakyThrows;
-import org.stellar.anchor.api.exception.SepException;
 import org.stellar.anchor.api.platform.GetTransactionResponse;
 import org.stellar.anchor.api.platform.PlatformTransactionData;
 import org.stellar.anchor.api.sep.AssetInfo;
 import org.stellar.anchor.api.sep.SepTransactionStatus;
-import org.stellar.anchor.api.shared.*;
+import org.stellar.anchor.api.shared.Amount;
+import org.stellar.anchor.api.shared.RefundPayment;
+import org.stellar.anchor.api.shared.Refunds;
 import org.stellar.anchor.asset.AssetService;
-import org.stellar.anchor.platform.data.JdbcSep24Transaction;
-import org.stellar.anchor.platform.data.JdbcSep31Transaction;
-import org.stellar.anchor.platform.data.JdbcSepTransaction;
 import org.stellar.anchor.sep24.Sep24RefundPayment;
 import org.stellar.anchor.sep24.Sep24Refunds;
+import org.stellar.anchor.sep24.Sep24Transaction;
 import org.stellar.anchor.sep31.Sep31Refunds;
+import org.stellar.anchor.sep31.Sep31Transaction;
 
 public class TransactionHelper {
-  @SneakyThrows
-  public static GetTransactionResponse toGetTransactionResponse(
-      JdbcSepTransaction txn, AssetService assetService) {
-    switch (txn.getProtocol()) {
-      case "24":
-        return toGetTransactionResponse((JdbcSep24Transaction) txn, assetService);
-      case "31":
-        return toGetTransactionResponse((JdbcSep31Transaction) txn);
-      default:
-        throw new SepException(String.format("Unsupported protocol:%s", txn.getProtocol()));
-    }
-  }
-
-  static GetTransactionResponse toGetTransactionResponse(JdbcSep31Transaction txn) {
+  public static GetTransactionResponse toGetTransactionResponse(Sep31Transaction txn) {
     Refunds refunds = null;
     if (txn.getRefunds() != null) {
       refunds = toRefunds(txn.getRefunds(), txn.getAmountInAsset());
@@ -61,8 +47,8 @@ public class TransactionHelper {
         .build();
   }
 
-  static GetTransactionResponse toGetTransactionResponse(
-      JdbcSep24Transaction txn, AssetService assetService) {
+  public static GetTransactionResponse toGetTransactionResponse(
+      Sep24Transaction txn, AssetService assetService) {
     Refunds refunds = null;
     if (txn.getRefunds() != null) {
       refunds = toRefunds(txn.getRefunds(), txn.getAmountInAsset());
@@ -86,7 +72,6 @@ public class TransactionHelper {
         .startedAt(txn.getStartedAt())
         .updatedAt(txn.getUpdatedAt())
         .completedAt(txn.getCompletedAt())
-        .transferReceivedAt(txn.getTransferReceivedAt())
         .message(txn.getMessage())
         .refunds(refunds)
         .stellarTransactions(txn.getStellarTransactions())
@@ -100,7 +85,7 @@ public class TransactionHelper {
   }
 
   private static String makeAsset(
-      @Nullable String dbAsset, AssetService service, JdbcSep24Transaction txn) {
+      @Nullable String dbAsset, AssetService service, Sep24Transaction txn) {
     if (dbAsset != null) {
       return dbAsset;
     }
