+++ platform/src/main/java/org/stellar/anchor/platform/event/DefaultEventService.java
@@ -13,10 +13,9 @@ import org.stellar.anchor.asset.AssetService;
 import org.stellar.anchor.config.event.EventConfig;
 import org.stellar.anchor.event.EventPublisher;
 import org.stellar.anchor.event.EventService;
-import org.stellar.anchor.platform.data.JdbcSepTransaction;
-import org.stellar.anchor.platform.utils.TransactionHelper;
 import org.stellar.anchor.sep24.Sep24Transaction;
 import org.stellar.anchor.sep31.Sep31Transaction;
+import org.stellar.anchor.util.TransactionHelper;
 
 public class DefaultEventService implements EventService {
   private final EventConfig eventConfig;
@@ -34,13 +33,12 @@ public class DefaultEventService implements EventService {
   @Override
   @SneakyThrows
   public void publish(Sep24Transaction txn, AnchorEvent.Type type) {
-    JdbcSepTransaction jdbcTxn = (JdbcSepTransaction) txn;
     AnchorEvent event =
         AnchorEvent.builder()
             .id(UUID.randomUUID().toString())
             .sep("24")
             .type(TRANSACTION_CREATED)
-            .transaction(TransactionHelper.toGetTransactionResponse(jdbcTxn, assetService))
+            .transaction(TransactionHelper.toGetTransactionResponse(txn, assetService))
             .build();
     publish(event);
   }
@@ -48,13 +46,12 @@ public class DefaultEventService implements EventService {
   @Override
   @SneakyThrows
   public void publish(Sep31Transaction txn, AnchorEvent.Type type) {
-    JdbcSepTransaction jdbcTxn = (JdbcSepTransaction) txn;
     AnchorEvent event =
         AnchorEvent.builder()
             .id(UUID.randomUUID().toString())
             .sep("31")
             .type(type)
-            .transaction(TransactionHelper.toGetTransactionResponse(jdbcTxn, assetService))
+            .transaction(TransactionHelper.toGetTransactionResponse(txn))
             .build();
     publish(event);
   }
